build: add lib64 folder before lib at LDFLAGS
Fixes: https://github.com/analogdevicesinc/iio-oscilloscope/issues/72

It's been reported that on OpenSUSE, there is a build error due to the
arch incompatibility between libs installed in /usr/lib64 and /usr/lib.

This error, is not present on Debian/Ubuntu systems.

OpenSUSE may handle things differently with respect to multi-arch setups.

So, to handle this, the /usr/lib64 is added to the LDFLAGS search-path,
before the /usr/lib path.
This should handle the OpenSUSE case [and hopefully similar cases], and it
also works on Debian/Ubuntu.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>